+++ lib/core/extensions/scroll_controller.dart
@@ -7,7 +7,8 @@ extension ScrollControllerExtension on ScrollController {
     if (this.hasClients)
       this.animateTo(
         0.00,
-        duration: const Duration(milliseconds: LunaUI.ANIMATION_SPEED * 2),
+        duration:
+            const Duration(milliseconds: LunaUI.ANIMATION_SPEED_SCROLLING),
         curve: Curves.easeInOutQuart,
       );
   }