+++ lib/core/models/types.dart
@@ -1,4 +1,5 @@
 export 'types/browser.dart';
 export 'types/indexer_icon.dart';
 export 'types/links.dart';
+export 'types/list_view_option.dart';
 export 'types/loading_state.dart';