+++ lib/modules/settings/routes/configuration_radarr/pages/default_options.dart
@@ -3,10 +3,10 @@ import 'package:lunasea/core.dart';
 import 'package:lunasea/modules/radarr.dart';
 import 'package:lunasea/modules/settings.dart';
 
-class SettingsConfigurationRadarrDefaultSortingRouter
+class SettingsConfigurationRadarrDefaultOptionsRouter
     extends SettingsPageRouter {
-  SettingsConfigurationRadarrDefaultSortingRouter()
-      : super('/settings/configuration/radarr/sorting');
+  SettingsConfigurationRadarrDefaultOptionsRouter()
+      : super('/settings/configuration/radarr/options');
 
   @override
   Widget widget() => _Widget();
@@ -36,7 +36,7 @@ class _State extends State<_Widget> with LunaScrollControllerMixin {
 
   Widget _appBar() {
     return LunaAppBar(
-      title: 'Default Sorting & Filtering',
+      title: 'settings.DefaultOptions'.tr(),
       scrollControllers: [scrollController],
     );
   }
@@ -45,14 +45,51 @@ class _State extends State<_Widget> with LunaScrollControllerMixin {
     return LunaListView(
       controller: scrollController,
       children: [
+        LunaHeader(text: 'radarr.Movies'.tr()),
+        _filteringMovies(),
         _sortingMovies(),
         _sortingMoviesDirection(),
-        _filteringMovies(),
-        const LunaDivider(),
+        _viewMovies(),
+        LunaHeader(text: 'radarr.Releases'.tr()),
+        _filteringReleases(),
         _sortingReleases(),
         _sortingReleasesDirection(),
-        _filteringReleases(),
       ],
+    );
+  }
+
+  Widget _viewMovies() {
+    RadarrDatabaseValue _db = RadarrDatabaseValue.DEFAULT_VIEW_MOVIES;
+    return _db.listen(
+      builder: (context, box, _) {
+        LunaListViewOption _view = _db.data;
+        return LunaBlock(
+          title: 'lunasea.View'.tr(),
+          body: [TextSpan(text: _view.readable)],
+          trailing: const LunaIconButton.arrow(),
+          onTap: () async {
+            List<String> titles = LunaListViewOption.values
+                .map<String>((view) => view.readable)
+                .toList();
+            List<IconData> icons = LunaListViewOption.values
+                .map<IconData>((view) => view.icon)
+                .toList();
+
+            Tuple2<bool, int> values = await SettingsDialogs().setDefaultOption(
+              context,
+              title: 'lunasea.View'.tr(),
+              values: titles,
+              icons: icons,
+            );
+
+            if (values.item1) {
+              LunaListViewOption _opt = LunaListViewOption.values[values.item2];
+              context.read<RadarrState>().moviesViewType = _opt;
+              _db.put(_opt);
+            }
+          },
+        );
+      },
     );
   }
 
@@ -60,15 +97,22 @@ class _State extends State<_Widget> with LunaScrollControllerMixin {
     RadarrDatabaseValue _db = RadarrDatabaseValue.DEFAULT_SORTING_MOVIES;
     return _db.listen(
       builder: (context, box, _) => LunaBlock(
-        title: 'Movies Sort Category',
+        title: 'settings.SortCategory'.tr(),
         body: [TextSpan(text: (_db.data as RadarrMoviesSorting).readable)],
         trailing: const LunaIconButton.arrow(),
         onTap: () async {
           List<String> titles = RadarrMoviesSorting.values
               .map<String>((sorting) => sorting.readable)
               .toList();
-          Tuple2<bool, int> values = await RadarrDialogs()
-              .setDefaultSortingOrFiltering(context, titles: titles);
+          List<IconData> icons = List.filled(titles.length, LunaIcons.SORT);
+
+          Tuple2<bool, int> values = await SettingsDialogs().setDefaultOption(
+            context,
+            title: 'settings.SortCategory'.tr(),
+            values: titles,
+            icons: icons,
+          );
+
           if (values.item1) {
             _db.put(RadarrMoviesSorting.values[values.item2]);
             context.read<RadarrState>().moviesSortType = _db.data;
@@ -85,7 +129,7 @@ class _State extends State<_Widget> with LunaScrollControllerMixin {
         RadarrDatabaseValue.DEFAULT_SORTING_MOVIES_ASCENDING;
     return _db.listen(
       builder: (context, box, _) => LunaBlock(
-        title: 'Movies Sort Direction',
+        title: 'settings.SortDirection'.tr(),
         body: [
           TextSpan(
             text:
@@ -109,15 +153,22 @@ class _State extends State<_Widget> with LunaScrollControllerMixin {
     RadarrDatabaseValue _db = RadarrDatabaseValue.DEFAULT_FILTERING_MOVIES;
     return _db.listen(
       builder: (context, box, _) => LunaBlock(
-        title: 'Movies Filter Category',
+        title: 'settings.FilterCategory'.tr(),
         body: [TextSpan(text: (_db.data as RadarrMoviesFilter).readable)],
         trailing: const LunaIconButton.arrow(),
         onTap: () async {
           List<String> titles = RadarrMoviesFilter.values
               .map<String>((filter) => filter.readable)
               .toList();
-          Tuple2<bool, int> values = await RadarrDialogs()
-              .setDefaultSortingOrFiltering(context, titles: titles);
+          List<IconData> icons = List.filled(titles.length, LunaIcons.FILTER);
+
+          Tuple2<bool, int> values = await SettingsDialogs().setDefaultOption(
+            context,
+            title: 'settings.FilterCategory'.tr(),
+            values: titles,
+            icons: icons,
+          );
+
           if (values.item1) {
             _db.put(RadarrMoviesFilter.values[values.item2]);
             context.read<RadarrState>().moviesFilterType = _db.data;
@@ -131,16 +182,25 @@ class _State extends State<_Widget> with LunaScrollControllerMixin {
     RadarrDatabaseValue _db = RadarrDatabaseValue.DEFAULT_SORTING_RELEASES;
     return _db.listen(
       builder: (context, box, _) => LunaBlock(
-        title: 'Releases Sort Category',
+        title: 'settings.SortCategory'.tr(),
         body: [TextSpan(text: (_db.data as RadarrReleasesSorting).readable)],
         trailing: const LunaIconButton.arrow(),
         onTap: () async {
           List<String> titles = RadarrReleasesSorting.values
               .map<String>((sorting) => sorting.readable)
               .toList();
-          Tuple2<bool, int> values = await RadarrDialogs()
-              .setDefaultSortingOrFiltering(context, titles: titles);
-          if (values.item1) _db.put(RadarrReleasesSorting.values[values.item2]);
+          List<IconData> icons = List.filled(titles.length, LunaIcons.SORT);
+
+          Tuple2<bool, int> values = await SettingsDialogs().setDefaultOption(
+            context,
+            title: 'settings.SortCategory'.tr(),
+            values: titles,
+            icons: icons,
+          );
+
+          if (values.item1) {
+            _db.put(RadarrReleasesSorting.values[values.item2]);
+          }
         },
       ),
     );
@@ -151,7 +211,7 @@ class _State extends State<_Widget> with LunaScrollControllerMixin {
         RadarrDatabaseValue.DEFAULT_SORTING_RELEASES_ASCENDING;
     return _db.listen(
       builder: (context, box, _) => LunaBlock(
-        title: 'Releases Sort Direction',
+        title: 'settings.SortDirection'.tr(),
         body: [
           TextSpan(
             text:
@@ -170,16 +230,25 @@ class _State extends State<_Widget> with LunaScrollControllerMixin {
     RadarrDatabaseValue _db = RadarrDatabaseValue.DEFAULT_FILTERING_RELEASES;
     return _db.listen(
       builder: (context, box, _) => LunaBlock(
-        title: 'Releases Filter Category',
+        title: 'settings.FilterCategory'.tr(),
         body: [TextSpan(text: (_db.data as RadarrReleasesFilter).readable)],
         trailing: const LunaIconButton.arrow(),
         onTap: () async {
           List<String> titles = RadarrReleasesFilter.values
-              .map<String>((filter) => filter.readable)
+              .map<String>((sorting) => sorting.readable)
               .toList();
-          Tuple2<bool, int> values = await RadarrDialogs()
-              .setDefaultSortingOrFiltering(context, titles: titles);
-          if (values.item1) _db.put(RadarrReleasesFilter.values[values.item2]);
+          List<IconData> icons = List.filled(titles.length, LunaIcons.FILTER);
+
+          Tuple2<bool, int> values = await SettingsDialogs().setDefaultOption(
+            context,
+            title: 'settings.FilterCategory'.tr(),
+            values: titles,
+            icons: icons,
+          );
+
+          if (values.item1) {
+            _db.put(RadarrReleasesFilter.values[values.item2]);
+          }
         },
       ),
     );